+++ Assets/Runtime/Core/Models/Execution.cs
@@ -114,7 +114,7 @@ namespace Appwrite.Models
             logs: map["logs"].ToString(),
             errors: map["errors"].ToString(),
             duration: Convert.ToDouble(map["duration"]),
-            scheduledAt: map["scheduledAt"]?.ToString()
+            scheduledAt: map.ContainsKey("scheduledAt") ? map["scheduledAt"]?.ToString() : null
         );
 
         public Dictionary<string, object?> ToMap() => new Dictionary<string, object?>()